Split process.go and process_test.go into multiple files
